设备类，事件类，handle类，桥接类，信息汇总到界面显示
1、设备类与设备通信。单独一个线程
2、取得信息后，解析信息后，构建事件，通过QCoreApplication::postEvent，发送事件。
3、Handle类中注册，接收目标。重写customEvent，解析设备信息。
3、设备信息结构体化，Q_GADGET，Q_DECLARE_METATYPE(Devicedata)。只用申明一个Q_PROPERTY。
4、设备类取用，在一个全局设备单例管理类中获得。

+++ PracticeQml/DeviceMonitor/SimulatedDeviceEvent.h
@@ -23,9 +23,9 @@ public:
 
 public:
     int device_id;
-    int voltage;
-    int current;
-    int soc;
+    double voltage;
+    double current;
+    double soc;
     QString version;
